feat: add version update check and banner notification
Introduce a feature to check for application updates by fetching a version JSON file. If a new version is detected, a banner notification is displayed, prompting the user to reload the page to update. The check runs on mount and every 5 minutes thereafter. This ensures users are always using the latest version of the application.